pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "circulatory-autogen"
version = "0.1.0"
description = "Generation and calibration of CellML circulatory system models from module/vessel arrays"
readme = "README.md"
requires-python = ">=3.7"
license = {text = "Apache-2.0"}
authors = [
    {name = "Finbar J. Argus"}
]
keywords = ["cellml", "circulatory", "modeling", "parameter-identification"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

dependencies = [
    "numpy>=1.19.0",
    "pandas>=1.3.0",
    "pyyaml>=5.4.0",
    "mpi4py>=3.0.0",
    "matplotlib>=3.3.0",
    "scipy>=1.7.0",
    "emcee>=3.0.0",
    "corner>=2.2.0",
    "SALib==1.5.1",
    "seaborn>=0.11.0",
    "rdflib>=5.0.0",
    "ruamel.yaml>=0.17.0",
    "sympy>=1.8.0",
    "tqdm>=4.62.0",
    "statsmodels>=0.12.0",
    "schwimmbad>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "pytest-mpi>=0.6.0",
    "pytest-xdist>=2.5.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
]

[tool.pytest.ini_options]
# Pytest configuration
# NOTE: Tests should be run using the OpenCOR Python shell.
# Use ./run_pytest.sh to run tests, which automatically uses the Python
# from opencor_pythonshell_path.sh
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "mpi: marks tests that require MPI",
    "need_autogen: waits for all autogeneration tests to finish before running",
    "autogen_task: marks an autogeneration test that signals completion",
    "autogen_rank(idx): autogeneration test assigned rank index",
    "xdist_group(name): group marker used by pytest-xdist to serialize groups",
]
# Coverage options
[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.black]
line-length = 100
target-version = ['py37', 'py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | generated_models
)/
'''

[tool.mypy]
python_version = "3.7"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

